[zsh] Split zprofile into zshenv
zshenv does not source zprofile anymore.  PATHs and stuff are set in
zshenv now, but the locale is kept in zprofile.

Consequently, the locale is not forcibly set in non-interactive sessions
anymore, because that broke some programs, see
https://github.com/bbatsov/projectile/issues/487

We now only set the locale in login shells, to cover SSH and friends.
In OS X, Terminal.app already gives us a ready locale environment.